Multiplication bug fixed & some FFT speed up

+++ Coursework/number/number.h
@@ -5,14 +5,14 @@
 
 class Number {
  private:
-  std::vector<uint16_t> digits;
+  std::vector<int> digits;
   bool negative = false;
 
  public:
   Number();
   Number(const std::string &s);
   Number(const int64_t &n);
-  Number(const std::vector<uint16_t> &v, bool n = false);
+  Number(const std::vector<int> &v, bool n = false);
 
   size_t size() const;
   bool is_negative() const;
